refactor: Establish client info structure
