+++ Cargo.toml
@@ -11,3 +11,4 @@ serde = { version = "1.0.210", features = ["derive"] }
 serde_json = "1.0.131"
 tokio = { version = "1.40.0", features = ["full"] }
 bytes = "1.0"
+uuid = { version = "1.11.0", features = ["v4", "serde"] }
