+++ common/Cargo.toml
@@ -9,3 +9,4 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 bytes = { workspace = true }
 tokio = { workspace = true }
+uuid = { workspace = true }
