+++ common/src/fs.rs
@@ -1,13 +1,11 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct DirTree {
-    files: Vec<FileMetadata>,
+#[derive(Debug, Serialize, Deserialize)]
+pub struct File {
+    name: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FileMetadata {
-    name: String,
-    size: u64,
-    modified: u64,
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FileMap {
+    files: Vec<File>,
 }
