+++ common/src/lib.rs
@@ -1,10 +1,9 @@
+pub mod client;
 pub mod event;
 pub mod fs;
-pub mod peer;
-pub mod server;
 pub mod tcp;
 pub mod util;
 
-pub use fs::{DirTree, FileMetadata};
-pub use server::{ClientInfo, Server, ServerStatus};
+pub use client::{Client, Peer};
+pub use fs::{File, FileMap};
 pub use tcp::{listener::Listener, message, transmitter::Transmitter};
