--- common/src/peer.rs
@@ -1,4 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Peer {}
